@@ -1,4 +1,4 @@
-class RemoveNickNameFromReviews < ActiveRecord::Migration
+class RemoveNickNameFromReviews < ActiveRecord::Migration[4.2]
   def change
     remove_column :reviews, :nickname, :string
   end
